refactor(reload): move resource types and reload decision structs to dedicated files and remove redundant code
